MTN: move decorator to run_engine
